refactor and rename response_models
