+++ main.py
@@ -4,47 +4,49 @@ from typing import Annotated
 from fastapi import BackgroundTasks, FastAPI, Form, HTTPException, status, UploadFile
 from fastapi.responses import StreamingResponse
 
+from polysoleval.models import *
 from polysoleval.datafile import validate
 from polysoleval.evaluate import evaluate_dataset
 from polysoleval.globals import *
-from polysoleval.response_models import *
+from polysoleval import responses
 
 
 @asynccontextmanager
 async def lifespan(app: FastAPI):
     # load model descriptions
-    types = ModelType.all_from_yaml(MODELPATH / "model_types.yaml")
-
-    global MODEL_TYPES
-    MODEL_TYPES = {m.name: m for m in types}
+    global NEURALNET_TYPES
+    NEURALNET_TYPES = {
+        m.name: m
+        for m in NeuralNetType.all_from_yaml(NEURALNETPATH / "model_types.yaml")
+    }
 
     # load ML models
-    for bg_model_path in MODELPATH.glob("*-Bg.pt"):
-        model_name, range_name, _ = bg_model_path.stem.split("-")
+    for bg_net_path in NEURALNETPATH.glob("*-Bg.pt"):
+        net_name, range_name, _ = bg_net_path.stem.split("-")
 
-        model_type = MODEL_TYPES.get(model_name, None)
-        if model_type is None:
+        net_type = NEURALNET_TYPES.get(net_name, None)
+        if net_type is None:
             continue
 
-        bth_model_path = bg_model_path.with_stem(f"{model_name}-{range_name}-Bth")
-        if not bth_model_path.is_file():
+        bth_net_path = bg_net_path.with_stem(f"{net_name}-{range_name}-Bth")
+        if not bth_net_path.is_file():
             continue
 
         range_path = RANGEPATH / f"{range_name}.yaml"
         if not range_path.is_file():
             continue
-        MODEL_INSTANCES[(model_name, range_name)] = ModelInstance(
-            model_type=model_type,
+        NEURALNET_PAIRS[NetRangePairNames(net_name, range_name)] = NeuralNetPair(
+            neuralnet_type=net_type,
             range_path=range_path,
-            bg_model_path=bg_model_path,
-            bth_model_path=bth_model_path,
+            bg_net_path=bg_net_path,
+            bth_net_path=bth_net_path,
         )
 
     yield
 
     # clear ML models
-    MODEL_TYPES.clear()
-    MODEL_INSTANCES.clear()
+    NEURALNET_TYPES.clear()
+    NEURALNET_PAIRS.clear()
 
 
 app = FastAPI(lifespan=lifespan)
@@ -56,17 +58,17 @@ def root():
 
 
 @app.get("/models")
-def get_models() -> ModelTypesResponse:
+def get_models() -> responses.NeuralNetTypes:
     """Return every model type.
 
     Returns:
         _type_: _description_
     """
-    return ModelTypesResponse(model_types=list(MODEL_TYPES.values()))
+    return responses.NeuralNetTypes(neuralnet_types=list(NEURALNET_TYPES.values()))
 
 
-@app.get("/models/{model_type}")
-def get_model_instances(model_type: str) -> ModelInstancesResponse:
+@app.get("/models/{neuralnet_name}")
+def get_model_instances(neuralnet_name: str) -> responses.ValidRangeSets:
     """Return all model instances of the given model type.
 
     Args:
@@ -75,13 +77,15 @@ def get_model_instances(model_type: str) -> ModelInstancesResponse:
     Returns:
         _type_: _description_
     """
-    instances = [v.range_set for k, v in MODEL_INSTANCES.items() if k[0] == model_type]
-    if instances:
-        return ModelInstancesResponse(model_instances=instances)
+    valid_sets = [
+        v.range_set for k, v in NEURALNET_PAIRS.items() if k.net_name == neuralnet_name
+    ]
+    if valid_sets:
+        return responses.ValidRangeSets(range_sets=valid_sets)
 
     raise HTTPException(
         status.HTTP_404_NOT_FOUND,
-        detail="No valid instances found for the chosen model type",
+        detail="No pre-trained networks found for the chosen neural net type",
     )
 
 
@@ -93,9 +97,9 @@ async def post_evaluate(
     mass: Annotated[float, Form(gt=0.0)],
     datafile: UploadFile,
     background_tasks: BackgroundTasks,
-) -> EvaluationResponse:
+) -> responses.Evaluation:
     # Validate first
-    instance = MODEL_INSTANCES.get((ml_model_name, range_name), None)
+    instance = NEURALNET_PAIRS.get(NetRangePairNames(ml_model_name, range_name), None)
     if instance is None:
         raise HTTPException(
             status.HTTP_400_BAD_REQUEST,
@@ -116,8 +120,8 @@ async def post_evaluate(
             mw,
             visc,
             rep_unit,
-            instance.bg_model,
-            instance.bth_model,
+            instance.bg_net,
+            instance.bth_net,
             instance.range_set,
         )
     except Exception as re:
@@ -132,12 +136,15 @@ async def post_evaluate(
 
 
 @app.get(
-    "/datafile/{token}", responses={200: {"content": {"text/plain; charset=utf-8": {}}}}
+    "/datafile/{token}",
+    responses={200: {"content": {"text/plain; charset=utf-8": {}}}},
 )
 def get_datafile(token: str) -> StreamingResponse:
     try:
         file_generator = HANDLER.get_generator(token)
     except KeyError as ke:
-        raise HTTPException(status.HTTP_404_NOT_FOUND, detail="data not found") from ke
+        raise HTTPException(
+            status.HTTP_404_NOT_FOUND, detail="datafile not found"
+        ) from ke
 
     return StreamingResponse(file_generator(), media_type="text/plain")
